Remove full text highlighting
